LeaveAutomation/requirements.txt
python-dotenv==1.0.1
imapclient==3.0.1
python-dateutil==2.9.0
requests==2.32.3
